move delete_cast_op_pass
