+++ paddle/fluid/inference/api/paddle_pass_builder.cc
@@ -276,6 +276,7 @@ GpuPassStrategy::GpuPassStrategy() : PassStrategy({}) {
         "transpose_flatten_concat_fuse_pass",  //
         "conv2d_fusion_layout_transfer_pass",  //
         "auto_mixed_precision_pass",           //
+        "delete_cast_op_pass",                 //
         "inplace_op_var_pass",                 // should be the last pass.
   });
